Fix all production audit findings from exhaustive review
Critical fixes:
- Fix dbOpener race condition with RWMutex protection for thread-safe
  concurrent access to the global DB driver
- Add recursion depth limits (256) to expression parser and evaluator
  to prevent stack overflow from pathologically nested expressions
- Fix goroutine leak in concurrent dispatcher: acquire semaphore before
  WaitGroup.Add to prevent orphaned goroutines on context cancellation
- Re-panic with original value (not PanicError wrapper) to preserve
  stack trace for debugging

High-severity fixes:
- Handle rand.Int() errors in random_string and random_int instead of
  silently using zero values when crypto/rand fails
- Return int64 from now_unix() to prevent overflow on 32-bit systems;
  update toInt/toFloat to handle int64 type
- Use context.Background() for fleet teardown so cleanup completes
  even when the execution context is cancelled

Medium-severity fixes:
- Validate that concurrent mode requires concurrency > 0 (previously
  silently defaulted to 1)
- Fleet setup failure now returns detailed error in SuiteResult with
  step-level error information for debugging
- Wire fleet registry in CLI (run.go) so fleet providers are available

Low-severity fixes:
- Fix parser doc comment to match implementation (allows literals and
  variable references, not just function calls)
- Add .gitignore entry for coverage output files

+++ core/expression/builtins_test.go
@@ -433,7 +433,7 @@ func TestNowUnix(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	n := val.(int)
+	n := val.(int64)
 	if n < 1000000000 {
 		t.Errorf("too small: %d", n)
 	}
